# bank-lending-frontend/src/app/services/models/credit-request-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CreditRequestResponse {
  amount?: number;
  date?: string;
  descriptionStatus?: string;
  firstName?: string;
  id?: number;
  lastName?: string;
  status?: 'APPROVED' | 'REFUSED' | 'UNDER_CONSIDERATION' | 'SENT_FOR_REVISION';
  surName?: string;
  type?: 'HIRED_WORKERS' | 'SELF_EMPLOYED' | 'INDIVIDUAL_ENTREPRENEUR' | 'PENSIONER';
  userId?: number;
}