# bank-lending-frontend/src/app/services/models/user-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UserResponse {
  addressFact?: string;
  addressRegister?: string;
  dateOfBirth?: string;
  email?: string;
  firstName?: string;
  id?: number;
  lastname?: string;
  passportSerialNumber?: string;
  phoneNumber?: string;
  surName?: string;
}