# bank-lending-frontend/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification",
    "description": "OpenApi documentation for Spring Security",
    "contact": {
      "name": "Timur Sharf",
      "url": "https://github.com/xzcPozer",
      "email": "timurkasharf@gmail.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088/api/v1",
      "description": "Local ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/credit-request/change/by/solvency/{creditRequestId}": {
      "put": {
        "tags": [
          "CreditRequest"
        ],
        "operationId": "changeCreditRequest",
        "parameters": [
          {
            "name": "creditRequestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "solvency"
                ],
                "type": "object",
                "properties": {
                  "solvency": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/credit-request/change/by/solvency/and/employment/{creditRequestId}": {
      "put": {
        "tags": [
          "CreditRequest"
        ],
        "operationId": "changeCreditRequest_1",
        "parameters": [
          {
            "name": "creditRequestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "employment",
                  "solvency"
                ],
                "type": "object",
                "properties": {
                  "solvency": {
                    "type": "string",
                    "format": "binary"
                  },
                  "employment": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/check": {
      "post": {
        "tags": [
          "User"
        ],
        "operationId": "checkUserData",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/credit-request/add": {
      "post": {
        "tags": [
          "CreditRequest"
        ],
        "operationId": "addCreditRequest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "requestStr",
                  "solvency"
                ],
                "type": "object",
                "properties": {
                  "requestStr": {
                    "type": "string"
                  },
                  "solvency": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/credit-request/add/by/hired-worker": {
      "post": {
        "tags": [
          "CreditRequest"
        ],
        "operationId": "addCreditRequest_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "employment",
                  "requestStr",
                  "solvency"
                ],
                "type": "object",
                "properties": {
                  "requestStr": {
                    "type": "string"
                  },
                  "solvency": {
                    "type": "string",
                    "format": "binary"
                  },
                  "employment": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/{userId}": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "getSelectedUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/user/get/by/serial-number": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "getUserBySerialNumber",
        "parameters": [
          {
            "name": "serialNum",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/user/all-client": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "getAllClient",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseUserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/user/all-client/by/sort": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "getAllClientBySort",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "isAsc",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseUserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/credit-request/request/by/{creditRequestId}": {
      "get": {
        "tags": [
          "CreditRequest"
        ],
        "operationId": "getCreditRequestById",
        "parameters": [
          {
            "name": "creditRequestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditRequestResponse"
                }
              }
            }
          }
        }
      }
    },
    "/credit-request/my-requests": {
      "get": {
        "tags": [
          "CreditRequest"
        ],
        "operationId": "getAll",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseCreditRequestClientResponse"
                }
              }
            }
          }
        }
      }
    },
    "/credit-request/all-requests": {
      "get": {
        "tags": [
          "CreditRequest"
        ],
        "operationId": "getAllF",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseCreditRequestResponse"
                }
              }
            }
          }
        }
      }
    },
    "/credit-request/all-requests/with/lending-officer": {
      "get": {
        "tags": [
          "CreditRequest"
        ],
        "operationId": "getAllWithLendingOfficer",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseCreditRequestResponseForDirector"
                }
              }
            }
          }
        }
      }
    },
    "/credit-request/all-requests/by/processed": {
      "get": {
        "tags": [
          "CreditRequest"
        ],
        "operationId": "getAll_1",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "isProcessed",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseCreditRequestResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserRequest": {
        "required": [
          "addressFact",
          "addressRegister",
          "dateOfBirth",
          "email",
          "firstName",
          "lastname",
          "passportSerialNumber",
          "phoneNumber"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "surName": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "passportSerialNumber": {
            "type": "string"
          },
          "addressRegister": {
            "type": "string"
          },
          "addressFact": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "surName": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "passportSerialNumber": {
            "type": "string"
          },
          "addressRegister": {
            "type": "string"
          },
          "addressFact": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "PageResponseUserResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "CreditRequestResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "surName": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "APPROVED",
              "REFUSED",
              "UNDER_CONSIDERATION",
              "SENT_FOR_REVISION"
            ]
          },
          "type": {
            "type": "string",
            "enum": [
              "HIRED_WORKERS",
              "SELF_EMPLOYED",
              "INDIVIDUAL_ENTREPRENEUR",
              "PENSIONER"
            ]
          },
          "amount": {
            "type": "integer",
            "format": "int32"
          },
          "descriptionStatus": {
            "type": "string"
          }
        }
      },
      "CreditRequestClientResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "loanPurpose": {
            "type": "string"
          },
          "sum": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "enum": [
              "APPROVED",
              "REFUSED",
              "UNDER_CONSIDERATION",
              "SENT_FOR_REVISION"
            ]
          },
          "descriptionStatus": {
            "type": "string"
          }
        }
      },
      "PageResponseCreditRequestClientResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreditRequestClientResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "PageResponseCreditRequestResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreditRequestResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "CreditRequestResponseForDirector": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "clientFirstName": {
            "type": "string"
          },
          "clientLastName": {
            "type": "string"
          },
          "clientSurName": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "APPROVED",
              "REFUSED",
              "UNDER_CONSIDERATION",
              "SENT_FOR_REVISION"
            ]
          },
          "lendingOfficerFirstName": {
            "type": "string"
          },
          "lendingOfficerLastName": {
            "type": "string"
          },
          "lendingOfficerSurName": {
            "type": "string"
          }
        }
      },
      "PageResponseCreditRequestResponseForDirector": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreditRequestResponseForDirector"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth with bearer token",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}